+++ packages/erebus-sdk/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @erebus-sh/sdk
 
+## 0.0.119
+
+### Patch Changes
+
+- Update the README.md
+- update build config, to remove console and debugger on production build
+
+## 0.0.118
+
+### Patch Changes
+
+- Update the README.md
+- update build config, to remove console and debugger on production build
+
 ## 0.0.117
 
 ### Patch Changes